feat: share GPU tier classifier via IPC
