+++ src/renderer/src/lib/telemetry.ts
@@ -75,23 +75,3 @@ export function toModelDirectoryBucket(directory: string | undefined): string {
 export function isTerminalModelDownloadStatus(status: ModelDownloadStatus): boolean {
   return status === 'completed' || status === 'error' || status === 'cancelled'
 }
-
-/** Coarse hardware tier for cohort filtering; raw gpu_model / gpu_vram_gb are also sent for finer slicing. */
-export type GpuTier = 'high' | 'mid' | 'low' | 'sub_low' | 'apple' | 'cpu_only'
-
-export function deriveGpuTier(opts: {
-  vendor: string | null | undefined
-  vramGb: number | null | undefined
-}): GpuTier {
-  const vendor = (opts.vendor ?? '').toLowerCase()
-  // Main reports 'mps' for Apple Silicon; also accept literal 'apple'.
-  if (vendor === 'apple' || vendor === 'mps') return 'apple'
-  const vram = opts.vramGb ?? 0
-  if (!vendor) return 'cpu_only'
-  if (vendor !== 'nvidia' && vendor !== 'amd') return 'sub_low'
-  if (vram <= 0) return 'cpu_only'
-  if (vram >= 24) return 'high'
-  if (vram >= 12) return 'mid'
-  if (vram >= 6) return 'low'
-  return 'sub_low'
-}
